requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-querycount==0.8.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
